agents: models, media and style namespaces run on the capability registry (PR 4 of tool-class retirement)
Ports find_model, list_models, list_provider_models, the seven
provider-agnostic media tools, the three VLM judges, and the taste pair
to capability modules. The providers map and the optional LongTermMemory
binding become run fields; the old classes survive as deprecated thin
CapabilityTool subclasses with unchanged constructors. UNGATED moves
from mcp-tools.ts into capabilities/invoke.ts.
